backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
pandas==2.2.3
numpy==1.25.2
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20